+++ assets/js/test1.js
@@ -1,4 +1,6 @@
 function setDataUrlToBlobLink(base64DataUrl, mimeType, fileName, linkId) {
+    if (base64DataUrl == '') { return; }
+
     // 1. Extract only pure data part from Base64
     const byteCharacters = atob(base64DataUrl.split(',')[1]);
     const byteNumbers = new Array(byteCharacters.length);
@@ -33,3 +35,29 @@ function setDataUrlToBlobLink(base64DataUrl, mimeType, fileName, linkId) {
 }
 // dummy call for minify
 setDataUrlToBlobLink('','','','');
+
+function getAnchorsDownloadHref(htmlId) {
+    if (htmlId == '') { return; }
+    const elm = document.getElementById(htmlId);
+    if (elm) {
+        return {download: elm.download, href: elm.href};
+    }
+    return {download: '', href: ''};
+}
+getAnchorsDownloadHref('');
+
+function parseBlobData_dxsend(url) {
+    if (url == '') { return; }
+    let xhr=new XMLHttpRequest();
+    xhr.open('GET',url,true);
+    xhr.responseType='blob';
+    xhr.onload=function(e) {
+        let reader=new FileReader();
+        reader.readAsDataURL(this.response);
+        reader.onloaded=function(){
+            let base64data=reader.result;
+            dioxus.send(base64data);
+        }
+    }
+}
+parseBlobData('');