+++ src/components/test1.rs
@@ -1,7 +1,26 @@
 use dioxus::prelude::*;
 
+#[cfg(feature = "desktop")]
+use anyhow::Result;
+#[cfg(feature = "desktop")]
+use serde::{Deserialize, Serialize};
+
 const APP_PNG: Asset = asset!("/assets/app.png");
 const TEST1_JS: Asset = asset!("/assets/js/test1.js", AssetOptions::js().with_minify(true));
+
+#[cfg(feature = "desktop")]
+#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
+pub struct AnchorInfo {
+    pub download: Option<String>,
+    pub href: Option<String>,
+}
+#[cfg(feature = "desktop")]
+impl AnchorInfo {
+    pub fn from_json_str(s: &str) -> Result<Self> {
+        let r = serde_json::from_str(s)?;
+        Ok(r)
+    }
+}
 
 #[component]
 pub fn Test1() -> Element {
@@ -50,18 +69,47 @@ pub fn Test1() -> Element {
             }
             br {}
             // download link: asset file
-            a { id: "lnk1", download: "app.png", href: APP_PNG, "Download Link: asset file" }
+            a {
+                id: "lnk1",
+                onclick: move |evt| async move {
+                    download_file("lnk1").await;
+                    evt.stop_propagation();
+                },
+                download: "app.png",
+                href: APP_PNG,
+                "Download Link: asset file"
+            }
             br {}
             // download link: data
             a {
                 id: "lnk2",
+                onclick: move |evt| async move {
+                    download_file("lnk2").await;
+                    evt.stop_propagation();
+                },
                 download: "{dl_data_fnm}",
                 href: "{dl_data_src}",
                 "{dl_data_msg}"
             }
             br {}
             // download link: blob
-            a { id: "lnk3", "{dl_blob_msg}" }
+            a {
+                id: "lnk3",
+                /*
+                onclick: move |evt| async move {
+                    download_file("lnk3").await;
+                    evt.stop_propagation();
+                },
+                */
+                target: "_blank",
+                onclick: move |evt| {
+                    evt.stop_propagation();
+                    spawn(async move{
+                        download_file("lnk3").await;
+                    });
+                },
+                "{dl_blob_msg}"
+            }
         }
     }
 }
@@ -142,4 +190,48 @@ async fn input_file_onchange_(
             }
         }
     }
+}
+
+#[cfg(not(feature = "desktop"))]
+async fn download_file(_id: &str) {}
+
+#[cfg(feature = "desktop")]
+async fn download_file(id: &str) {
+    //dioxus_logger::tracing::debug!("data: {:?}", evt.data());
+    let js = format!(r#"{{return getAnchorsDownloadHref('{}');}}"#, id);
+    let v = document::eval(&js).await.unwrap();
+    let s = v.to_string();
+    let anchorinfo = AnchorInfo::from_json_str(&s).unwrap();
+    let filename = anchorinfo.download.unwrap();
+    dioxus_logger::tracing::debug!("filename: {filename}");
+    if let Some(path) = rfd::FileDialog::new().set_file_name(filename).save_file() {
+        let content = anchorinfo.href.unwrap();
+        let is_data = content.starts_with("data:");
+        let is_blob = content.starts_with("blob:");
+        if is_data || is_blob {
+            let data_url = if is_blob {
+                let js = format!(r#"{{parseBlobData_dxsend('{}');}}"#, content);
+                let mut eval = document::eval(&js);
+                let data_url = eval.recv::<String>().await.unwrap();
+                data_url
+            } else {
+                content
+            };
+            save_data_uri0(&data_url, &path).unwrap();
+        }
+    }
+}
+
+#[cfg(feature = "desktop")]
+fn save_data_uri0(url: &str, path: &std::path::PathBuf) -> Result<()> {
+    use base64::{engine::general_purpose::STANDARD_NO_PAD, Engine as _};
+    // data uri schema:
+    //   data:[<mediatype>][;base64],<data>
+    let base64_s = url.split_once(',').unwrap().1;
+    let v = STANDARD_NO_PAD.decode(base64_s)?;
+    if let Ok(mut file) = std::fs::File::create(path) {
+        use std::io::Write;
+        let _ = file.write_all(&v);
+    }
+    Ok(())
 }