+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* desktop download
 * `android-versionCode-inc` into Makefile
 * download link: blob
 * download link: data